fix coments
